Unified reactive and async clients
