--- RocketApp/Domain/Sources/RocketsClient/ClientAsync/RocketsClientAsync+test.swift
@@ -1,11 +1,0 @@
-#if DEBUG
-import Dependencies
-import Foundation
-
-public extension RocketsClientAsync {
-   static let test = Self(
-    getRocket: unimplemented("\(Self.self).getRocket"),
-    getAllRockets: unimplemented("\(Self.self).getAllRockets")
-  )
-}
-#endif
