+++ RocketApp/Domain/Sources/RocketsClient/RocketsClient+dependencyManaging.swift
@@ -2,7 +2,7 @@ import Dependencies
 import Foundation
 
 enum RocketsClientKey: DependencyKey {
-  public static let liveValue: RocketsClient = .live
+  public static let liveValue: RocketsClient = .liveReact
   #if DEBUG
   public static let testValue: RocketsClient = .test
   #endif
