+++ RocketApp/Domain/Sources/RocketsClient/RocketsClient+preview.swift
@@ -1,6 +1,7 @@
+import Combine
 import Foundation
 
-public extension RocketsClientAsync {
+public extension RocketsClient {
   static let preview = Self(
     getRocket: { _ in RocketDetail.mock },
     getAllRockets: { [RocketDetail.mock] }
