+++ RocketApp/Domain/Sources/RocketsClient/RocketsClient+test.swift
