+++ RocketApp/Domain/Sources/RocketsClient/RocketsClient.swift
@@ -1,6 +1,7 @@
+import Combine
 import Foundation
 
-public struct RocketsClientAsync {
+public struct RocketsClient {
   public typealias GetRocketFunction = (String) async throws -> RocketDetail
   public typealias GetAllRocketsFunction = () async throws -> [RocketDetail]
