+++ RocketApp/Domain/Sources/RocketsClient/RocketsClientAsyncError.swift
