+++ RocketApp/Domain/Sources/RocketsClient/RocketsClientError.swift
