+++ RocketApp/Domain/Sources/RocketsClient/URLs.swift
@@ -1,5 +1,5 @@
 import Foundation
 
 public enum URLs {
-  public static let baseURL = "https://api.spacexdata.com"
+  public static let baseURL = "https://api.spaxdata.com"
 }
